Move transformer into module
